Fix warning about fabs in GameViewController.m

+++ skeleton/RacerGame/RacerGame/GameViewController.m
@@ -518,8 +518,8 @@
 
 - (void)update
 {
-    float aspect = fabsf(self.view.bounds.size.width /
-                         self.view.bounds.size.height);
+    float aspect = fabs(self.view.bounds.size.width /
+                        self.view.bounds.size.height);
     GLKMatrix4 projectionMatrix = GLKMatrix4MakePerspective(GLKMathDegreesToRadians(50.0f), aspect, 0.1f, 100.0f);
     self.effect.transform.projectionMatrix = projectionMatrix;
